requirements.txt
python-telegram-bot==13.14
apscheduler==3.8.1
python-dotenv==0.19.2